requirements.txt
psycopg[binary]>=3.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
fastapi>=0.111.1
uvicorn>=0.23.0
pydantic>=2.6.0
httpx>=0.27.0
geoalchemy2>=0.14.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
